graças a deus

+++ src/functions/eventData.c
@@ -3,7 +3,7 @@
 #include <stdlib.h>
 
 struct Event {  // Dynamic list
-    int date;
+    long int date;
     int cardCode;
     int gateCode;
     char eventType;
@@ -18,7 +18,7 @@ struct Event {  // Dynamic list
     @param eventType: The state of the event (E/S)
     @return: the new struct event
 */
-struct Event *newEvent(int date, int cardCode, int gateCode, char eventType) {
+struct Event *newEvent(long int date, int cardCode, int gateCode, char eventType) {
     struct Event *event = malloc(sizeof(struct Event));  // Creat a event with the correct quantitative of bits
 
     event->date = date;
